+++ apps/backend/src/services/teams.ts
@@ -10,7 +10,6 @@ import { InferUIMessageChunk, readUIMessageStream } from 'ai';
 import { Card, Chat, Message, SentMessage, Thread } from 'chat';
 
 import { generateChartImage } from '../components/generate-chart';
-import { generateMapImage } from '../components/generate-map';
 import * as chartImageQueries from '../queries/chart-image';
 import * as chatQueries from '../queries/chat.queries';
 import * as feedbackQueries from '../queries/feedback.queries';
@@ -31,6 +30,7 @@ import {
 	createTextBlock,
 	EXCLUDED_TOOLS,
 	formatMessagingError,
+	renderMapImage,
 } from '../utils/messaging-provider';
 import { agentService } from './agent';
 import { posthog, PostHogEvent } from './posthog';
@@ -408,7 +408,10 @@ class TeamsService {
 			return;
 		}
 		state.renderedToolCallIds.add(part.toolCallId);
-		const png = await this._renderMapImage(part, state, ctx);
+		const png = await renderMapImage(part, state, this._projectId, {
+			chatId: ctx.chatId,
+			toolCallId: part.toolCallId,
+		});
 		if (!png) {
 			await this._pushMapLinkCard(part, ctx);
 			return;
@@ -425,30 +428,6 @@ class TeamsService {
 				context: { chatId: ctx.chatId, toolCallId: part.toolCallId },
 			});
 			await this._pushMapLinkCard(part, ctx);
-		}
-	}
-
-	private async _renderMapImage(
-		part: Extract<UIMessagePart, { type: 'tool-display_map' }>,
-		state: StreamState,
-		ctx: ConversationContext,
-	): Promise<Buffer | null> {
-		if (part.state !== 'output-available') {
-			return null;
-		}
-		const sqlOutput = state.sqlOutputs.get(part.input.query_id);
-		if (!sqlOutput) {
-			return null;
-		}
-		try {
-			const customBoundaries = await projectQueries.getCustomBoundaries(this._projectId);
-			return await generateMapImage({ config: part.input, rows: sqlOutput.rows, customBoundaries });
-		} catch (error) {
-			logger.error(`Map image generation failed: ${String(error)}`, {
-				source: 'system',
-				context: { chatId: ctx.chatId, toolCallId: part.toolCallId },
-			});
-			return null;
 		}
 	}
